requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
jinja2==3.1.3
chromadb==0.4.22
tiktoken==0.5.2
python-multipart==0.0.6
chromadb
fastapi
httpx
jinja2
openai
passlib[bcrypt]
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
python-multipart
sqlalchemy
tiktoken
uvicorn[standard]
psycopg2-binary
email-validator
numpy<2.0
bcrypt==4.0.1
